Anti_Inspection/mobile-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
google-generativeai
python-dotenv
